exception hamdling

+++ Exercicios/ecommerce/src/main/java/com/ecommerceJava/ecommerce/services/ProductService.java
@@ -4,15 +4,16 @@ package com.ecommerceJava.ecommerce.services;
 import com.ecommerceJava.ecommerce.dto.ProductDTO;
 import com.ecommerceJava.ecommerce.entities.Product;
 import com.ecommerceJava.ecommerce.repositories.ProductRepository;
+import com.ecommerceJava.ecommerce.services.exeptions.DatabaseException;
+import com.ecommerceJava.ecommerce.services.exeptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
-import java.util.Optional;
-import java.util.stream.Collectors;
 
 @Service
 public class ProductService {
@@ -22,10 +23,9 @@ public class ProductService {
 
     @Transactional(readOnly = true)
     public ProductDTO findById(long id) {
-        Optional<Product> result  = repository.findById(id);
-        Product product = result.get();
-        ProductDTO dto = new ProductDTO(product);
-        return dto;
+        Product product = repository.findById(id).orElseThrow(
+                ()-> new ResourceNotFoundException("recurso nao encontrado"));
+        return new ProductDTO(product); // Lancando a propria execcao do error
 
     }
 
@@ -58,16 +58,37 @@ public class ProductService {
 
     @Transactional()
     public ProductDTO update (Long id,ProductDTO dto) {
-        Product entity = repository.getReferenceById(id); // pega a referencia do id que eu passar, sem ir no BD
-        copyDtoToEntity(dto,entity);
-        return new ProductDTO(entity);
+        try {
+            Product entity = repository.getReferenceById(id); // pega a referencia do id que eu passar, sem ir no BD
+            copyDtoToEntity(dto,entity);
+            return new ProductDTO(entity);
+
+        }
+        catch (EntityNotFoundException e){
+            throw new ResourceNotFoundException("recurso nao encontrado");
+        }
 
     }
 
-    @Transactional
-    public void  delete (long id) {
-        repository.deleteById(id);
+    @Transactional(propagation = Propagation.SUPPORTS) // so vai executar se o metodo estiver no contexto de outra transacao.
+    public void delete(Long id) {
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+
+        try {
+            repository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
+
+// delete antigo.
+//    @Transactional
+//    public void  delete (long id) {
+//        repository.deleteById(id);
+//    }
 
 
     private void copyDtoToEntity(ProductDTO dto, Product entity) {
